# doc_api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkillForge API",
    "version": "0.0.1",
    "description": "API REST de la plateforme SkillForge : gestion des modules de formation, scenarios, et assets (upload via SFTP).\n\n**Authentification** : la plupart des endpoints requierent un JWT (header `Authorization: Bearer <token>`). Cliquez sur le bouton « Authorize » dans Swagger UI et collez le token (sans le prefixe « Bearer »).\n\n**Note** : les controllers sont mappes a la racine (ex. `/modules`), il n'y a pas de context-path `/api`. Le webhook SFTP est authentifie par le header `X-Webhook-Secret`, pas par JWT.",
    "contact": {
      "name": "Equipe SkillForge",
      "email": "ngapepouekarl@gmail.com"
    },
    "license": {
      "name": "Proprietaire"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Environnement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/scenarios/{scenarioId}/assets": {
      "get": {
        "tags": ["Scenarios <-> Assets"],
        "summary": "Lister les assets d'un scenario",
        "description": "Liste les assets ACTIVE d'un scenario. Acces : TRAINEE, TRAINER, ADMIN.",
        "operationId": "getAssetsByScenario",
        "parameters": [ { "$ref": "#/components/parameters/PathScenarioId" } ],
        "responses": {
          "200": {
            "description": "Liste des assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AssetSummaryResponse" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/scenarios/{scenarioId}/assets/{assetId}": {
      "post": {
        "tags": ["Scenarios <-> Assets"],
        "summary": "Attacher un asset a un scenario",
        "description": "Associe un asset existant a un scenario. Acces : ADMIN, TRAINER.",
        "operationId": "attachAssetToScenario",
        "parameters": [
          { "$ref": "#/components/parameters/PathScenarioId" },
          { "$ref": "#/components/parameters/PathAssetId" }
        ],
        "responses": {
          "200": { "description": "Asset attache" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Scenarios <-> Assets"],
        "summary": "Detacher un asset d'un scenario",
        "description": "Retire l'association entre un asset et un scenario. Acces : ADMIN, TRAINER.",
        "operationId": "detachAssetFromScenario",
        "parameters": [
          { "$ref": "#/components/parameters/PathScenarioId" },
          { "$ref": "#/components/parameters/PathAssetId" }
        ],
        "responses": {
          "204": { "description": "Asset detache" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/assets/{assetId}/scenarios": {
      "get": {
        "tags": ["Scenarios <-> Assets"],
        "summary": "Lister les scenarios utilisant un asset",
        "description": "Liste les scenarios qui referencent un asset donne. Acces : ADMIN, TRAINER.",
        "operationId": "getScenariosByAsset",
        "parameters": [ { "$ref": "#/components/parameters/PathAssetId" } ],
        "responses": {
          "200": {
            "description": "Liste des scenarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ScenarioSummaryResponse" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentification JWT. Saisissez uniquement le token (sans « Bearer »)."
      }
    },
    "parameters": {
...
      "PathScenarioId": {
        "name": "scenarioId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "PathAssetId": {
        "name": "assetId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Non authentifie (JWT manquant ou invalide)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Acces refuse (role insuffisant)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Ressource introuvable",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ScenarioSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "moduleId": { "type": "string", "format": "uuid" },
          "active": { "type": "boolean" }
        }
      },
...
      "AssetSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "type": { "type": "string" },
          "fileUrl": { "type": "string", "nullable": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "status": { "type": "integer", "format": "int32" },
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": {
            "type": "object",
            "nullable": true,
            "additionalProperties": { "type": "string" },
            "description": "Erreurs de validation par champ"
          }
        }
      },
...
    }
  }
}